+++ png2c/png2c.py
@@ -12,17 +12,17 @@ def png2c(path, name):
 
     img = img.reshape((-1, 4))
 
-    rgb888 = [(a << 24) | (r << 16) | (g << 8) | b for (r, g, b, a) in img]
+    rgb888 = [(a << 24) | (r << 16) | (g << 8) | b for (r, g, b, a) in img.astype(np.uint32)]
 
     with open(f'{name}_rgb888.h', 'w') as f:
         f.write(f'const uint32_t Image_{name}_rgb888[{height}][{width}] = {{\n')
         for i, x in enumerate(rgb888):
             if i % 16 == 0 and i != 0:
                 f.write('\n')
-            f.write('0x%08X, ' %x.astype(np.uint32))
+            f.write('0x%08X, ' %x)
         f.write('\n};')
     
-    rgb565 = [(a << 16) | ((r >> 3) << 11) | ((g >> 2) << 5) | (b >> 3) for (r, g, b, a) in img]
+    rgb565 = [(a << 16) | ((r >> 3) << 11) | ((g >> 2) << 5) | (b >> 3) for (r, g, b, a) in img.astype(np.uint32)]
     
     with open(f'{name}_rgb565.h', 'w') as f:
         f.write(f'const uint32_t Image_{name}_rgb565[{height}][{width}] = {{\n')
